+++ Runtime/VisualScripting.Flow/Framework/Time/Cooldown.cs
@@ -141,12 +141,7 @@ namespace Unity.VisualScripting
             data.isListening = true;
         }
 
-        void IGraphEventListener.StopListening(GraphStack stack, bool destroyed)
-            => StopListening(stack, destroyed);
-
-        public void StopListening(GraphStack stack) => StopListening(stack, true);
-
-        private void StopListening(GraphStack stack, bool destroyed)
+        public void StopListening(GraphStack stack)
         {
             var data = stack.GetElementData<Data>(this);