+++ Runtime/VisualScripting.Flow/ScriptMachine.cs
@@ -41,7 +41,7 @@ namespace Unity.VisualScripting
 
             if (hasGraph)
             {
-                graph.StopListening(reference, true);
+                graph.StopListening(reference);
             }
         }
 
@@ -51,16 +51,7 @@ namespace Unity.VisualScripting
 
             if (hasGraph)
             {
-                graph.StopListening(reference, false);
-            }
-        }
-
-        protected override void OnDestroy()
-        {
-            base.OnDestroy();
-            if (hasGraph)
-            {
-                graph.StopListening(reference, true);
+                graph.StopListening(reference);
             }
         }