+++ Runtime/VisualScripting.Flow/SubgraphUnit.cs
@@ -146,18 +146,13 @@ namespace Unity.VisualScripting
             stack.GetElementData<Data>(this).isListening = true;
         }
 
-        void IGraphEventListener.StopListening(GraphStack stack, bool destroyed)
-            => StopListening(stack, destroyed);
-
-        public void StopListening(GraphStack stack) => StopListening(stack, true);
-
-        private void StopListening(GraphStack stack, bool destroyed)
+        public void StopListening(GraphStack stack)
         {
             stack.GetElementData<Data>(this).isListening = false;
 
             if (stack.TryEnterParentElement(this))
             {
-                (nest.graph as IGraphEventListener).StopListening(stack, destroyed);
+                nest.graph.StopListening(stack);
                 stack.ExitParentElement();
             }
         }