+++ Runtime/VisualScripting.Flow/Unit.cs
@@ -66,7 +66,7 @@ namespace Unity.VisualScripting
         {
             if (this is IGraphEventListener listener)
             {
-                listener.StopListening(instance, true);
+                listener.StopListening(instance);
             }
 
             base.Uninstantiate(instance);