+++ Runtime/VisualScripting.State/StateTransition.cs
@@ -56,7 +56,7 @@ namespace Unity.VisualScripting
         {
             if (this is IGraphEventListener listener)
             {
-                listener.StopListening(instance, true);
+                listener.StopListening(instance);
             }
 
             base.Uninstantiate(instance);