+++ Runtime/VisualScripting.State/SuperState.cs
@@ -44,16 +44,11 @@ namespace Unity.VisualScripting
             }
         }
 
-        void IGraphEventListener.StopListening(GraphStack stack, bool destroyed)
-            => StopListening(stack, destroyed);
-
-        public void StopListening(GraphStack stack) => StopListening(stack, true);
-
-        private void StopListening(GraphStack stack, bool destroyed)
+        public void StopListening(GraphStack stack)
         {
             if (stack.TryEnterParentElement(this))
             {
-                (nest.graph as IGraphEventListener).StopListening(stack, destroyed);
+                nest.graph.StopListening(stack);
                 stack.ExitParentElement();
             }
         }